Updated timepicker with better default values
Based on my experience, the default date range wasn't used if the default was absolute or since mode in the timepicker, especially if user visited the dashboard before. This will set the default "from" time to be whatever is in the configuration of the timepicker on subsequent banana visits as well.