APOLLO-14091 New log dashboards for Fusion 4.1 release (#323)
* Prevent duplicate filters from being added to Filter panel.
Now a user will not see duplicate filters being added to the Filter panel,
when clicking on the same filter in a chart for multiple times.

* Add Sub-row feature to Table panel.
Sub-row allows a specified field to be shown by itself in its own row for
each event in the table. You can also specify the maximum number of characters
to be shown and the character offset number (from the beginning of the line).

* Fix RangeFacet panel so that it does not add a global filter when first initialized.
Now when a range facet panel is added to a dashboard, it will initially
use the panel's settings, instead of creating a new range filter that
effects the whole dashboard.

* Allow duplicate for time filters so that Histogram panel will work correctly.
Change default collection from 'logs' to 'system_logs' in config.js for Fusion 4.1.

* Add three new dashboard json files: Service Logs, Access Logs, and Combined Logs.
Service Logs will be the new home (default) dashboard in Fusion.

* Release version 1.6.20
Update README.
Update dist files (from grunt build).